+++ wallet/src/ui/app/pages/home/index.tsx
@@ -65,5 +65,6 @@ export { default as TokensPage } from './tokens';
 export { default as TransactionDetailsPage } from './transaction-details';
 export { default as TransactionsPage } from './transactions';
 export { default as TransferCoinPage } from './transfer-coin';
+export { default as TransferNFTPage } from './transfer-nft';
 export { default as NFTDetialsPage } from './nft-detials';
 export { default as StakeNew } from './stake-new';
