+++ wallet/src/ui/app/pages/home/transfer-nft/validation.ts
